allow optout of session keys
